Extract the leaf-compaction pipeline into a mixin (WS5 Seam 6)
Move the compaction gate + pipeline out of LCMEngine into a CompactionMixin in
the new compaction.py, and mix it in AHEAD of ContextEngine
(`class LCMEngine(CompactionMixin, ContextEngine)`).

Moved verbatim (byte-identical bodies): should_compress, should_compress_preflight
(public), _replay_diff_requests_ingest_cleanup, _has_ignored_backlog_outside_fresh_tail,
_leaf_compaction_candidate_status, _working_leaf_chunk_tokens, _select_oldest_leaf_chunk,
and the main compress entry point. They keep running bound to the engine instance,
so `_ingest_cursor`, `_store`, `_dag`, `_lifecycle`, the per-turn caches and the
~60 sibling helpers stay on the engine and resolve via self.

Two deliberate boundary choices:
- The mixin is placed BEFORE ContextEngine in the bases so its compress /
  should_compress / should_compress_preflight override the ContextEngine
  protocol defaults (which also define those names).
- The summarize-with-rescue sub-cluster (_summarize_leaf_chunk_with_rescue and
  its _next_leaf_rescue_chunk / _is_retry_worthy_leaf_summary_error helpers)
  stays on the engine, because ~40 tests patch the module-level
  summarize_with_escalation on hermes_lcm.engine to intercept the LLM call.
  compress() reaches it via self as a callback-out, so behaviour is unchanged and
  no test is touched.

engine.py shrinks by ~752 lines. Drops engine.py's now-exclusive
`_contains_sensitive_redaction` import (moved to compaction.py).

Behaviour-preserving refactor (WS5 engine decomposition). Rollback: git revert
this commit re-inlines the methods; no data or state migration involved.
